+++ CHANGELOG.md
@@ -4,6 +4,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### 🔄 Changed
+
+# [0.2.0](https://github.com/GetStream/stream-video-swift/releases/tag/0.2.0)
+_July 18, 2023_
+
 ### ✅ Added
 - SDK version info sent in all requests
 - Call participants shown in the lobby view